backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
python-dotenv>=1.0.0
aiohttp>=3.11.0
aiosqlite>=0.21.0

# Data & Analysis
yfinance>=0.2.50
pandas>=2.2.0
pandas-ta>=0.3.14b
numpy>=2.1.0

# GenAI & Agents
langgraph>=0.3.0
langchain-google-genai>=2.1.0
langchain-core>=0.3.0

# News
feedparser>=6.0.11
beautifulsoup4>=4.12.0

# Utilities
httpx>=0.28.0
cachetools>=5.5.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.23.0
google-generativeai>=0.8.0
